Bump SBPF to v0.11.0 (#5839)
* Bump to SBPF v0.11.0

* Cleanup - EbpfError::InvalidVirtualAddress

* Refactor - Return region index

* Refactor - Simplify CoW in MemoryMapping

* Inlines Serializer::push_account_data_region().

+++ transaction-context/src/lib.rs
@@ -498,7 +498,7 @@ impl TransactionContext {
     }
 
     /// Returns a new account data write access handler
-    pub fn account_data_write_access_handler(&self) -> Box<dyn Fn(u64) -> Result<u64, ()>> {
+    pub fn account_data_write_access_handler(&self) -> Box<dyn Fn(u32) -> Result<u64, ()>> {
         let accounts = Rc::clone(&self.accounts);
         Box::new(move |index_in_transaction| {
             // The two calls below can't really fail. If they fail because of a bug,
